+++ lib/command/minitest.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'command/minitest/assertions/assert_successful'
 require 'command/minitest/assertions/assert_valid'
 require 'command/minitest/assertions/assert_with_error'