+++ lib/command/minitest/assertions/assert_valid.rb
@@ -1,19 +1,3 @@
-module Minitest
-  module Assertions
-    def assert_valid(actual, *args)
-      assert valid?(actual, *args),
-             "Expected #{actual.inspect} to be valid?"
-    end
+# frozen_string_literal: true
 
-    def refute_valid(actual, *args)
-      refute valid?(actual, *args),
-             "Expected #{actual.inspect} to not be valid?"
-    end
-
-    private
-
-    def valid?(actual, *args)
-      actual.valid?(*args)
-    end
-  end
-end
+require 'commande/minitest/assertions/assert_valid'