+++ lib/commande/minitest.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'commande/minitest/assertions/assert_successful'
+require 'commande/minitest/assertions/assert_valid'
+require 'commande/minitest/assertions/assert_with_error'