+++ test/commande/errors_test.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require_relative '../test_helper'
 
-module Command
+module Commande
   class ErrorsTest < Minitest::Test
 
-    class CommandWithError
-      include Command
+    class CommandeWithError
+      include Commande
 
       def call
         error 'first'
@@ -14,22 +16,22 @@ module Command
     end
 
     def test_result_is_not_successful
-      result = CommandWithError.call
+      result = CommandeWithError.call
       refute_successful result
     end
 
     def test_result_has_first_error
-      result = CommandWithError.call
+      result = CommandeWithError.call
       assert_equal 'first', result.error
     end
 
     def test_result_has_fatal_error
-      result = CommandWithError.call
+      result = CommandeWithError.call
       assert_with_error 'fatal', result
     end
 
     def test_result_stopped_on_fatal_error
-      result = CommandWithError.call
+      result = CommandeWithError.call
       refute_with_error 'nope', result
     end
   end