+++ test/commande/outputs_test.rb
@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 require_relative '../test_helper'
 
-module Command
+module Commande
   class OutputsTest < Minitest::Test
 
-    class CommandWithOutput
-      include Command
+    class CommandeWithOutput
+      include Commande
 
       output :foo, :baz
 
@@ -19,21 +21,21 @@ module Command
     end
 
     def test_result_has_output
-      result = CommandWithOutput.call
+      result = CommandeWithOutput.call
 
       assert_respond_to(result, :foo)
       assert_equal 'foo', result.foo
     end
 
     def test_result_has_output_even_if_not_assigned
-      result = CommandWithOutput.call
+      result = CommandeWithOutput.call
 
       assert_respond_to(result, :baz)
       assert_nil result.baz
     end
 
     def test_result_does_not_have_other_instance_variables
-      result = CommandWithOutput.call
+      result = CommandeWithOutput.call
       refute_respond_to(result, :not_an_output)
     end
   end