+++ test/commande/transfer_test.rb
@@ -1,16 +1,18 @@
+# frozen_string_literal: true
+
 require_relative '../test_helper'
 
-module Command
+module Commande
   class TransferTest < Minitest::Test
 
-    class OuterCommand
-      include Command
+    class OuterCommande
+      include Commande
 
       output :foo, :bar
 
       def call(inner_args)
         self.foo = 'foo'
-        transfer! FailOnStringCommand.call(inner_args)
+        transfer! FailOnStringCommande.call(inner_args)
         self.bar = 'bar'
       end
 
@@ -19,8 +21,8 @@ module Command
       attr_accessor :foo, :bar
     end
 
-    class FailOnStringCommand
-      include Command
+    class FailOnStringCommande
+      include Commande
 
       output :inner_output
 
@@ -35,15 +37,15 @@ module Command
     end
 
     def test_result_matches_inner_successful
-      result = OuterCommand.call('this is an error')
+      result = OuterCommande.call('this is an error')
       refute_successful result
 
-      result = OuterCommand.call(42)
+      result = OuterCommande.call(42)
       assert_successful result
     end
 
     def test_transfer_merges_outputs
-      result = OuterCommand.call(42)
+      result = OuterCommande.call(42)
       assert_respond_to result, :foo
       assert_equal 'foo', result.foo
 
@@ -52,13 +54,13 @@ module Command
     end
 
     def test_failed_transfer_ends_call
-      result = OuterCommand.call('this is an error')
+      result = OuterCommande.call('this is an error')
       assert_respond_to result, :bar
       assert_nil result.bar
     end
 
     def test_transfer_merges_errors
-      result = OuterCommand.call('this is an error')
+      result = OuterCommande.call('this is an error')
       assert_equal result.error, 'this is an error'
     end
   end