+++ test/commande/version_test.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 require_relative '../test_helper'
 
-module Command
+module Commande
   class VersionTest < Minitest::Test
     def test_that_it_has_a_version_number
-      refute_nil ::Command::VERSION
+      refute_nil ::Commande::VERSION
     end
   end
 end