+++ test/test_helper.rb
@@ -1,5 +1,7 @@
+# frozen_string_literal: true
+
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
-require 'command'
-require 'command/minitest'
+require 'commande'
+require 'commande/minitest'
 
 require 'minitest/autorun'