+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0
+- Naming gem as `commande`.
+- Github require continues to work because `command` can be required.
+- Aliases using `command/**` for all `commande/**`
+
 ## 0.2.1
 
 - `assert_valid` and `refute_valid` take `*args`, to match `Command#call` (and thus `Command#valid?`)